Stunngtash
I apologize in advance for being completely blind to the internet policies and procedures. Not much is a secret around here so this is my attempt to fight back and get a better online identity. 
If anybody wanted to help me, or point me in the right direction then feel free to reach out. Because at this point..i am..
Perfectly Imperfect,
S-tash